Add an option for a primaryKey
The option is true by default and the default primary key's attribute
name is `id`. If a string instead of true is passed, the string will be
used as attribute name. The library asserts that a primary key is unique.

The factories to use for testing defined in specHelper had to be changed
to not use a primaryKey as their tests don't test this behavior.